==> fifa-world-cup-2022-prediction-main/1.get_tables_groupstage-Copy1.ipynb <==
import pandas as pd
import numpy as np
import pickle
from string import ascii_uppercase as alphabet

from sqlalchemy import create_engine

engine = create_engine(
    "postgresql+psycopg2://postgres:admin@localhost:5432/wc22_development"
)
con22 = engine.connect()

sql = """
SELECT *
FROM standings"""
standings = pd.read_sql(sql, con22)
standings

sql = """
SELECT id, name, T.group 
FROM teams T"""
teams = pd.read_sql(sql, con22)
teams

sqlUpd = """
UPDATE teams
SET name = 'South Korea'
WHERE id = 32"""
rp = con22.execute(sqlUpd)
rp.rowcount

teams["name"] = teams["name"].str.strip()

df_merge = pd.merge(teams, standings, on='id', how='inner')
df_merge

# extracting all tables in website
all_tables = pd.read_html('https://en.wikipedia.org/wiki/2022_FIFA_World_Cup')

len(all_tables)

all_tables[9] # First Group

df_A = all_tables[9]  # Group A
df_A

df_A.rename(columns={df_A.columns[1]:'name'}, inplace=True)
df_A.pop('Qualification')
df_A

df_A.loc[(df_A.Pos == 4) ,['name']] = ['Qatar']
df_A

df_B = all_tables[16]  # Group B
df_B.rename(columns={df_B.columns[1]:'name'}, inplace=True)
df_B.pop('Qualification')
df_B

df_C = all_tables[23]  # Group C
df_C.rename(columns={df_C.columns[1]:'name'}, inplace=True)
df_C.pop('Qualification')
df_C

df_D = all_tables[30]  # Group D
df_D.rename(columns={df_D.columns[1]:'name'}, inplace=True)
df_D.pop('Qualification')
df_D

df_E = all_tables[37]  # Group E
df_E.rename(columns={df_E.columns[1]:'name'}, inplace=True)
df_E.pop('Qualification')
df_E

df_F = all_tables[44]  # Group F
df_F.rename(columns={df_F.columns[1]:'name'}, inplace=True)
df_F.pop('Qualification')
df_F

df_G = all_tables[51]  # Group G
df_G.rename(columns={df_G.columns[1]:'name'}, inplace=True)
df_G.pop('Qualification')
df_G

df_H = all_tables[58]  # Group H
df_H.rename(columns={df_H.columns[1]:'name'}, inplace=True)
df_H.pop('Qualification')
df_H

#df_fixture['home'] = df_fixture['home'].str.strip()
df_B["name"] = df_B["name"].str.strip()

df_B

df_B.loc[(df_B.Pos == 1) ,['name']] = ['England']
df_B

cols = 'team_id W D L GF GA Pts'.split()
cols

### Begin of Loop

df_merge_2 = pd.merge(df_merge, df_H, on = 'name', how='inner')
df_merge_2

df_merge_2[cols]

df_merge_2[cols].to_sql("temp_table", con=con22, index=False, if_exists="replace")

sql = """
SELECT *
FROM temp_table"""
df_tmp = pd.read_sql(sql, con22)
df_tmp

sqlUpd = """
    UPDATE standings AS f
    SET wins=t."W",draws=t."D",losses=t."L",goals_for= t."GF",goals_against=t."GA",points=t."Pts"
    FROM temp_table AS t
    WHERE f.team_id = t.team_id
"""
print(sqlUpd)

with engine.begin() as conn:     # TRANSACTION
    conn.execute(sqlUpd)

### End of Loop

# A -> H
# 9 -> 7*8 + 9 = 65
all_tables = pd.read_html('https://en.wikipedia.org/wiki/2022_FIFA_World_Cup')
for i in range(9,65,7): # 13 20 27
    print(i)
    df = all_tables[i]
    df.rename(columns={df.columns[1]:'Team'}, inplace=True)
    df.pop('Qualification')
#     print(df)

print(df)

print(alphabet)

all_tables = pd.read_html('https://en.wikipedia.org/wiki/2022_FIFA_World_Cup')

dict_table = {}
for letter, i in zip(alphabet, range(13,69,7)): # A=11, B=18, ...
    df = all_tables[i]
    df.rename(columns={df.columns[1]:'Team'}, inplace=True)
    df.pop('Qualification')
    dict_table[f'Group {letter}'] = df

dict_table.keys()

dict_table['Group F']

# Upload (..verify if uploaded correctly)
with open('dict_table', 'wb') as output:
    pickle.dump(dict_table, output)
